Extract the command-line logic into its own module
