+++ self-serve.ts
@@ -1,6 +1,8 @@
 // Deno Standard Library
-import { parseArgs } from "jsr:@std/cli"
 import { contentType } from "jsr:@std/media-types"
+
+// Modules
+import * as cli from './src/cli.ts'
 
 // ----
 // SELF
@@ -399,69 +401,20 @@ class Self {
 // MAIN
 // ----
 
-const VERSION = "v0.3.0"
-
-const DEFAULT_HOST = "localhost"
-const DEFAULT_PORT = "5327"
-
-const HELP_MESSAGE = `self-serve [directory] [options]
-
-self-serve is a super simple HTTP static file server
-
-Options:
-  -d, --dir     Directory to serve (default: current directory)
-  -a, --host    Host address to listen on (default: ${DEFAULT_HOST})
-  -p, --port    Port to listen on (default: ${DEFAULT_PORT})
- 
-  -h, --help    Show this help message
-  -v, --version Show version number
-`
-
-/** Parses command-line arguments and returns an object with the parsed values */
-function parseCommandLineArguments(args: string[]): { dir: string, host: string, port: number, version: boolean, help: boolean } {
-    const flags = parseArgs(args, {
-        string: ["dir", "host", "port"],
-        boolean: ["help", "version"],
-        alias: {
-            "help": "h",
-            "version": "v",
-            "dir": "d",
-            "host": "a",
-            "port": "p",
-        },
-        default: {
-            dir: Deno.cwd(),
-            host: DEFAULT_HOST,
-            port: DEFAULT_PORT,
-        },
-    })
-
-    // The first non-flag argument is the directory
-    const dir = flags._.length > 0 ? String(flags._[0]) : flags.dir
-
-    const port = Number(flags.port)
-    if (isNaN(port) || port < 1 || port > 65535) {
-        console.error(`Invalid port number: ${flags.port}`)
-        Deno.exit(1)
-    }
-
-    return { ...flags, dir, port }
-}
-
 /** The main entrypoint of the application */
 async function main() {
     // Parse the command-line arguments
-    const args = parseCommandLineArguments(Deno.args)
+    const args = cli.parse(Deno.args)
 
     // Show the help message if `-h` or `--help` command-line option was passed in
     if (args.help) {
-        console.info(HELP_MESSAGE)
+        cli.showHelp()
         return Deno.exit(0)
     }
 
     // Show the version number if `-v` or `--version` command-line option was passed in
     if (args.version) {
-        console.info(VERSION)
+        cli.showVersion()
         return Deno.exit(0)
     }
